+++ kernel/server/block.go
@@ -3,6 +3,7 @@ package server
 import (
 	"net/http"
 
+	"publish/auth"
 	"publish/client"
 	"publish/config"
 
@@ -11,17 +12,17 @@ import (
 
 func block(c *gin.Context) {
 	root_id := c.Param("id")
-	r, err := client.GetBlockDomByID(client.C.R(), root_id)
+	r, err := client.GetBlockDomByID(client.C.R(), root_id, 0)
 	r = client.Response(c, r, err)
 	if r == nil {
+		auth.StatusSiyuanServerError(c)
 		return
 	}
 	blocks := r.Data.(map[string]interface{})["blocks"].([]interface{})
 	switch {
 	case len(blocks) == 0:
-		c.String(http.StatusNotFound, "Block Not Found")
-	case len(blocks) > 1:
-		c.String(http.StatusInternalServerError, "More than one block found")
+		auth.StatusBlockNotFound(c)
+		return
 	default:
 		block := blocks[0].(map[string]interface{})
 		c.HTML(http.StatusOK, "block.html", gin.H{