+++ CHANGELOG.md
@@ -6,3 +6,7 @@
   - `theme=light`: 明亮主题模式 | Light theme mode.
   - `theme=dark`: 暗黑主题模式 | Dark theme mode.
   - `theme=auto`: 根据浏览器环境自动选择主题模式 | Auto select theme mode by browser environment.
+- 支持使用文档块属性 `custom-publish-access` 设置访问权限 | Support use document block attribute `custom-publish-access` to set access permission.
+  - `public`: 公开访问 | Public access.
+  - `private`: 私有访问 | Private access.
+  - `protected`: 受保护访问 | Protected access.